feat(ui): Move deleting and editing repo to a separate settings page
As the run list for a repository is no longer an overview page but
dedicated run list, remove the edit and delete buttons from that page and
move the functionality and the existing edit page to a new settings page.
